+++ src/features/tools/xlf/xlf-translationState.ts
@@ -1,0 +1,5 @@
+export enum translationState {
+    missingTranslation = 0,
+    needsWorkTranslation = 1,
+    translated = 10
+}